+++ CHANGELOG.rst
@@ -4,7 +4,11 @@ Changelog
 -------
 * Fix bug in ecda (convert mpz into int when not supported)
 * add coinactivity example script
+* add blockstats example script
+* Add Blocks class, which uses get_block_range
 * PR #272: correct blockchain virtual op batch calls (thanks to @crokkon)
+* PR #276: blockchain: get_account_reputations fix for first (thanks to @crokkon)
+* PR #287: beempy witnessproperties: fix interest rate options
 
 0.24.20
 -------